LayoutTests/accessibility/mac/range-for-contenteditable-newline-expected.txt

some
hello world
test

abc



def
This tests that when there are newline characters within text controls, we can get the correct text range from index and length.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS textSome is 'some'
PASS textHello is 'hello'
PASS textWorld is 'world'
PASS textTest is 'test'
PASS textareaTextHello is 'hello'
PASS textareaTextWorld is 'world'
PASS textDEF is 'def'
PASS successfullyParsed is true

TEST COMPLETE
